% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpostal_helpers.R
\name{check_libpostal_docker}
\alias{check_libpostal_docker}
\title{Test If Libpostal Is Running}
\usage{
check_libpostal_docker(image = "libpostal")
}
\arguments{
\item{imagename}{a regular expression to search for in the image name}
}
\value{
a logical value indicating whether libpostal is running or not
}
\description{
test to see if libpostal-rest docker image is running on the current machine.
}
\details{
currently Windows only
}
