% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpostal_parse_address.R
\name{parse_address}
\alias{parse_address}
\title{Parse Address with libpostal}
\usage{
parse_address(serveraddress = "192.168.99.100:8080/parser", address,
  collapsestreet = T, simplify = T)
}
\arguments{
\item{serveraddress}{the address of a libpostal_rest server parser, by default 192.168.99.100:8080/parser}

\item{address}{a character vector of addresses to parse. Each address should be a single item}

\item{collapsestreet}{a logical value indicating whether house number and road name should be collapsed to addr1}

\item{simplify}{a logical value indicating, if TRUE, that simple character vector should be returned, or if FALSE, a data frame should be returned}
}
\value{
if simplify = TRUE, a character vector of addresses parsed by libpostal. If simplify = F, a data frame with up to 8 columns including:
  * house_number - usually refers to the external (street-facing) building number.
  * unit - apartment or suite number
  * po_box - post office box: typically found in non-physical (mail-only) addresses
  * road - street name(s)
  * city - name of the city
  * st - state abbreviation
  * zip - postal code
}
\description{
parse an address with libpostal. This depends on a running instance of libpostal with a REST API. Using a Docker image of libpostal_rest is recommended. Make sure that your Docker VM has at least 2GB of RAM available.
}
\details{
Use `check_libpostal_docker()` to ensure that your docker image is running. If not, you can start it with `start_libpostal_docker()`.
}
