% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpostal_helpers.R
\name{prep_query}
\alias{prep_query}
\title{Prep REST query}
\usage{
prep_query(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a vector of queries passable to a libpostal_rest api via POST in the form of `{\"query\": \"123 Easy St, Buffalo NY\"}`
}
\description{
prep a character vector to be a query for a REST API.
}
