% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_usps.R
\name{query_usps}
\alias{query_usps}
\title{Query USPS API}
\usage{
query_usps(queryurl)
}
\arguments{
\item{queryurl}{a semi-human readable query URL to the USPS API}
}
\value{
a data frame with the columns
}
\description{
Query USPS API
}
