% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpostal_helpers.R
\name{start_libpostal_docker}
\alias{start_libpostal_docker}
\title{Start Libpostal Docker Instance}
\usage{
start_libpostal_docker(dockername = "amazing_curran")
}
\arguments{
\item{dockername}{the name of a libpostal-rest docker instance}
}
\description{
starts a docker instance on the current machine. Currently Windows only.
}
