% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps_validate.R
\name{usps_validate}
\alias{usps_validate}
\title{Validate a mailing address}
\usage{
usps_validate(username, aptnum = NULL, street, city, st, zip = NULL,
  zip4 = NULL, simplify = T)
}
\arguments{
\item{username}{your USPS username. Sign up at https://registration.shippingapis.com/}

\item{aptnum}{optional apartment number, ie "Apt 38"}

\item{street}{required street number and name, eg. "123 Easy St"}

\item{city}{required city name, eg. "Decatur"}

\item{st}{required state name, eg. "IL"}

\item{zip}{optional 5 digit zip code, eg "12345"}

\item{zip4}{optional 4 digit zip code addition, eg. "1234"}

\item{simplify}{a logical value determining whether the results should be returned as a simple character vector or, if FALSE, a multicolum data frame}
}
\value{
a data frame with friendly column names, including addr1 (street address), city, st, zip and zip4
}
\description{
validate a mailing address against the USPS database.
}
